Filter cgroup mountpoints using the rootfs directory value
There can be multiple cgroups mounted and this helps gosigar use the appropriate one. If we are expecting the cgroup to be mounted under the rootfs directory then we can filter out any that are not contained under that directory. This helps when running inside of docker on a RHEL6 based machine where on the host cgroup data is in /cgroup but inside the container the mountpoints are /sys/fs/cgroup.