Windows: Do not fail ProcState.Get when run by non-admin
When running under Windows as a non-administrator  user, ProcState.Get()
will fail with the error "OpenProcessToken failed for pid=XXX:
Permission denied".

This is a known issue under Windows: A non-admin won't be able to
access the token of most processes, regardless of privileges
(i.e. SeDebugPrivilege).

We should still allow ProcState.Get to return with a missing username
instead of failing completely.